修复单词拼写错误manger->manager，修复数据库链接bug，Logs类改名为Log，在service增加新的方法。！修改了表！

+++ src/team/skadi/rental/dao/PowerDao.java
@@ -20,7 +20,7 @@ public interface PowerDao {
 	 * @param gtLeft 剩余电量
 	 * @return 充电宝列表
 	 */
-	List<Power> findPowersByPowerLeft(double gtLeft);
+	List<Power> findPowersByPowerLeft(int gtLeft);
 
 	/**
 	 * 根据指定的id查找充电宝
